+++ assets/js/workflow-diagram/nodes/Node.tsx
@@ -94,7 +94,6 @@ const Node = ({
   const startInfo = data?.startInfo as
     | { started_at: string; startBy: string }
     | undefined;
-  const isErrorRun = runData?.exit_reason !== 'success';
   // TODO: remember triggers
   const didRun = data.isRun
     ? !!runData || (!!data?.startInfo && isTriggerNode)
@@ -113,6 +112,7 @@ const Node = ({
       className={`group ${didRun ? 'opacity-100' : 'opacity-30'}`}
       data-a-node
       data-id={id}
+      data-testid={type === 'trigger' ? `trigger-node-${id}` : `job-node-${id}`}
       data-valid-drop-target={
         data.isValidDropTarget !== undefined
           ? String(data.isValidDropTarget)